src/test/java/com/fbm/services/UserServiceTest.java
package com.fbm.services;

import com.fbm.domain.User;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

/**
 * Created by Mocart on 10-Jul-17.
 */
@RunWith(SpringRunner.class)
@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
class UserServiceTest {

    @Autowired
    private UserService userService;


    @Test
    void getAllUsers() {
        List<User> users = userService.getAllUsers();

        System.out.println("hello");
        System.out.println("hello");
        System.out.println("hello");
        System.out.println("hello");
        assertThat(users.size()).isEqualTo(3);
    }

    @Test
    void getUserById() {
    }

    @Test
    void saveUser() {
    }

}